restructure code for v0
