+++ pkg/registry/v0/dockerhub.go
