+++ pkg/registry/v0/registry.go
