+++ pkg/scanner/v0/interface.go
