+++ pkg/scanner/v0/trivy.go
